Refactor pbench/__init__.py
Split out md5sum, rename_tb_link, and quarantine into
utils.py since this is more logical and it could possibly
be reused.
