+++ server/bin/pbench-verify-backup-tarballs.py
@@ -46,7 +46,8 @@ import sys
 import tempfile
 from enum import Enum
 
-from pbench import BadConfig, PbenchConfig, get_pbench_logger, md5sum
+from pbench import BadConfig, PbenchConfig, get_pbench_logger
+from pbench.utils import md5sum
 from pbench.report import Report
 from s3backup import Entry, S3Config
